Rename `GenericIcon` to `UnifiedIcon` for clarity
