--- src/Themes/NexusMods.Themes.NexusFluentDark/Resources/GenericIconControlTheme.axaml
@@ -1,93 +1,0 @@
-<ResourceDictionary xmlns="https://github.com/avaloniaui"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:icon="clr-namespace:NexusMods.App.UI.Controls.GenericIcon;assembly=NexusMods.App.UI">
-
-        <Design.PreviewWith>
-
-        <StackPanel Orientation="Horizontal">
-            <StackPanel.Resources>
-                <icon:IconValue x:Key="ProjektankerIconValue">
-                    <icon:IconValue.ProjektankerIconValueSetter>
-                        <icon:ProjektankerIcon Value="mdi-cog" />
-                    </icon:IconValue.ProjektankerIconValueSetter>
-                </icon:IconValue>
-
-                <icon:IconValue x:Key="AvaloniaImageValue">
-                    <icon:IconValue.AvaloniaImageSetter>
-                        <icon:AvaloniaImage Image="avares://NexusMods.App.UI/Assets/DesignTime/cyberpunk_game.png"/>
-                    </icon:IconValue.AvaloniaImageSetter>
-                </icon:IconValue>
-
-                <icon:IconValue x:Key="AvaloniaSvgValue">
-                    <icon:IconValue.AvaloniaSvgSetter>
-                        <icon:AvaloniaSvg Path="avares://NexusMods.App.UI/Assets/Icons/add_circle_24px.svg"/>
-                    </icon:IconValue.AvaloniaSvgSetter>
-                </icon:IconValue>
-
-                <icon:IconValue x:Key="AvaloniaPathIconValue">
-                    <icon:IconValue.AvaloniaPathIconSetter>
-                        <icon:AvaloniaPathIcon>
-                            <icon:AvaloniaPathIcon.Data>
-                                <StreamGeometry>m4 12 1.41 1.41L11 7.83V20h2V7.83l5.58 5.59L20 12l-8-8-8 8z</StreamGeometry>
-                            </icon:AvaloniaPathIcon.Data>
-                        </icon:AvaloniaPathIcon>
-                    </icon:IconValue.AvaloniaPathIconSetter>
-                </icon:IconValue>
-            </StackPanel.Resources>
-
-            <StackPanel.Styles>
-                <Style Selector="icon|GenericIcon">
-                    <Setter Property="Size" Value="100"/>
-                </Style>
-
-                <Style Selector="icon|GenericIcon.icon">
-                    <Setter Property="Value" Value="{StaticResource ProjektankerIconValue}"/>
-                </Style>
-
-                <Style Selector="icon|GenericIcon.image">
-                    <Setter Property="Value" Value="{StaticResource AvaloniaImageValue}"/>
-                </Style>
-
-                <Style Selector="icon|GenericIcon.svg">
-                    <Setter Property="Value" Value="{StaticResource AvaloniaSvgValue}"/>
-                </Style>
-
-                <Style Selector="icon|GenericIcon.pathIcon">
-                    <Setter Property="Value" Value="{StaticResource AvaloniaPathIconValue}"/>
-                </Style>
-            </StackPanel.Styles>
-
-            <icon:GenericIcon Value="{StaticResource ProjektankerIconValue}" />
-            <icon:GenericIcon Classes="icon"/>
-
-            <icon:GenericIcon Value="{StaticResource AvaloniaImageValue}"/>
-            <icon:GenericIcon Classes="image"/>
-
-            <icon:GenericIcon Value="{StaticResource AvaloniaSvgValue}"/>
-            <icon:GenericIcon Classes="svg"/>
-
-            <icon:GenericIcon Value="{StaticResource AvaloniaPathIconValue}"/>
-            <icon:GenericIcon Classes="pathIcon"/>
-
-            <!-- with max size -->
-            <icon:GenericIcon Classes="image" Size="Infinity" MaxSize="90"/>
-        </StackPanel>
-
-    </Design.PreviewWith>
-
-    <ControlTheme x:Key="GenericIconTheme" TargetType="icon:GenericIcon">
-        <Setter Property="Template">
-            <ControlTemplate>
-                <ContentPresenter Name="PART_ContentPresenter"
-                                  Content="{TemplateBinding Content}"
-                                  ContentTemplate="{TemplateBinding ContentTemplate}"
-                                  Height="{TemplateBinding Size}"
-                                  Width="{TemplateBinding Size}"
-                                  FontSize="{TemplateBinding Size}"
-                                  MaxHeight="{TemplateBinding MaxHeight}"
-                                  MaxWidth="{TemplateBinding MaxWidth}"/>
-            </ControlTemplate>
-        </Setter>
-    </ControlTheme>
-</ResourceDictionary>
-
