+++ src/Themes/NexusMods.Themes.NexusFluentDark/Styles/Controls/Icons/IconsStyles.axaml
@@ -2,7 +2,7 @@
     xmlns="https://github.com/avaloniaui"
     xmlns:icons="clr-namespace:Projektanker.Icons.Avalonia;assembly=Projektanker.Icons.Avalonia"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:genericIcon="clr-namespace:NexusMods.App.UI.Controls.GenericIcon;assembly=NexusMods.App.UI">
+    xmlns:unifiedIcon="clr-namespace:NexusMods.App.UI.Controls.UnifiedIcon;assembly=NexusMods.App.UI">
 
     <!-- Preview -->
     <Design.PreviewWith>
@@ -74,8 +74,8 @@
 
 
     <!-- Style Definitions -->
-    <Style Selector="genericIcon|GenericIcon">
-        <Setter Property="Theme" Value="{StaticResource GenericIconTheme}" />
+    <Style Selector="unifiedIcon|UnifiedIcon">
+        <Setter Property="Theme" Value="{StaticResource UnifiedIconTheme}" />
     </Style>
 
     <!-- Foreground color classes -->
